Fix transparent color parsing

+++ Common/src/main/java/net/puffish/skillsmod/config/colors/ColorConfig.java
@@ -34,7 +34,7 @@ public record ColorConfig(int argb) {
 					}
 					case 8 -> {
 						return Result.success(new ColorConfig(
-								Integer.parseInt(string, 16)
+								(int) Long.parseLong(string, 16)
 						));
 					}
 					default -> { }
